Ticker refactor 3
Adding the globals that are used in ticker.
Since we're juggling on_tick event in and out, we'll need to initialize those along with event handler subscription. For that a new function will be created.
Actually, on unsubscription we would have to call an opposite of that function, but we already have global.ArtifactCollectors table destroyed explicitly, and apart from that there are only four numbers left dangling. Their values cant really cause problems later on and they're not notably space-consuming.
Also, I've come with a better way of distributing updates.

As a finishing remark I'd like to make a confession: all previous commits while bringing progam to a more robust state won't take much of effect on the times. 
Firstly, I believe it can now be easily seen, that most time is wasted in buildings module where there's a lot of heavyweight game api function calls. I'll see to that, but I can't make promises when.
Secondly, the most heavy load here is probably brought by find_entities_filtered() call. 50x50 square is quite a big area, and seeing how it only takes 1 item per second might make players decide to put a dozen of them in one place. So if you'd wish to further increase speed of this module consider following ways:
-Make collectors grab more items in a single act, this way the interval between collections might be increased and thus heavy load find_entities will be called less often.
-Reduce the area of search. It seems natural to build these collectors behind the defenses, so  it is likely that the most of area they cover is free of biter-drops by virtue of being free of biters. You might probably consider making them directional, so that they look up the area not centered around them.

Well, I guess that's all I have to say for now. Bye.